sensor_server/requirements.txt
# Python 3.7 or higher is required.
ruamel.yaml